✨ feat: generate sitemap.xml for GitHub Pages static export
- Add sitemap.xml generation script using tsx
- Create dedicated sitemap config to avoid Next.js font dependencies
- Move sitemap generation from pages to app router
- Update build script to generate sitemap after Next.js build
- Ensure sitemap.xml is generated in docs/ for GitHub Pages

This change ensures proper SEO by generating a sitemap.xml file during the static export build process. We maintain an app route for development convenience while generating a static sitemap.xml in the docs/ directory for GitHub Pages hosting. The API routes warning is expected and acceptable for our use case.
