+++ src/config/sitemap.ts
@@ -1,0 +1,4 @@
+import { BUILD } from './build';
+import { PAGES, SITE_URL } from './pages';
+
+export { BUILD, PAGES, SITE_URL };
