+++ src/utils/sitemap.ts
@@ -1,0 +1,20 @@
+import { PAGES, SITE_URL } from '@/config/sitemap';
+
+export function generateSiteMap() {
+  return `<?xml version="1.0" encoding="UTF-8"?>
+    <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+      ${Object.values(PAGES)
+        .map((page) => {
+          return `
+            <url>
+              <loc>${SITE_URL}${page.path}</loc>
+              <lastmod>${new Date().toISOString()}</lastmod>
+              <changefreq>daily</changefreq>
+              <priority>${page.path === '/' ? '1.0' : '0.8'}</priority>
+            </url>
+          `;
+        })
+        .join('')}
+    </urlset>
+  `;
+}
